Title Limiter Improvements
The limiter on Event Titles has been improved to allow button functionality. That is, when a user wants to see the full length of an event title that has exceed the maximum line limit, they will be able to press the new TextButton disguised as ellipsis to view it. This button only appears on events with such limit breaking titles and will immediately disappear after the title has been extended to its full length.

This also resulted in the Event Dialog Stateless Widget becoming a Stateful Widget so that the title changes could be seen in real time.