Read/Unread Icon Changed
During feedback, it was found that the previous icon for unread messages was unclear. After looking at icons for similar apps, I found that they use a simple dot to convey unread messages/emails/events etc. Therefore, I have changed it to a simple dot too. Sometimes, simple is better, and I hope that this is the case for these darn icons.