Resource Screen Buttons now adjusted
The Resource Screen Buttons besides the Dark/Light Mode Toggle have been made to adjust to the size of the text within instead of being a large fixed size within a SizedBox. This may not be a complete solution as the padding size may differ between devices of different resolution.